Inline `create_iface`
